content/canvas/test/test_size.attributes.default.html

<!DOCTYPE HTML>
<title>Canvas test: size.attributes.default</title>
<!-- Testing: Default width/height -->
<script src="/MochiKit/packed.js"></script>
<script src="/tests/SimpleTest/SimpleTest.js"></script>
<link rel="stylesheet" href="/tests/SimpleTest/test.css">
<body>
<canvas id="c" ><p class="fallback">FAIL (fallback content)</p></canvas>
<script>

SimpleTest.waitForExplicitFinish();
MochiKit.DOM.addLoadEvent(function () {

var canvas = document.getElementById('c');
var ctx = canvas.getContext('2d');

ok(canvas.width == 300, "canvas.width == 300");
ok(canvas.height == 150, "canvas.height == 150");
ok(!canvas.hasAttribute('width'), "!canvas.hasAttribute('width')");
ok(!canvas.hasAttribute('height'), "!canvas.hasAttribute('height')");

SimpleTest.finish();

});
</script>
